+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Release Notes for Read Time
 
+## 1.0.2 2025-01-29
+
+### Changed
+
+- Updated supported PHP versions
+- Base plugin class name
+- Pass Entry variable into saveReadTime function instead of Event 
+- Move field loop into its own function
+- Updated SuperTable function
+
+### Added
+
+- ReadTime content model
+- Code comments
+- Error log statement
+
+### Removed
+
+- Global twig extensions
+- Unused field handle array
+
 ## 1.0.1 2025-01-24
 
 ### Changed